# internal/pkg/docs/swagger.json
{
  "components": {
    "schemas": {
      "dto.CodeExchangeRequest": {
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "dto.EmailLoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "is_admin": {
            "type": "boolean"
          },
          "password": {
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "dto.LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "is_admin": {
            "type": "boolean"
          },
          "is_first_login": {
            "type": "boolean"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "dto.PasswordResetConfirmRequest": {
        "properties": {
          "password": {
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
      "dto.PasswordResetRequest": {
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "dto.PasswordResetTokenCheck": {
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
      "dto.PasswordUpdateRequest": {
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "required": [
          "new_password"
        ],
        "type": "object"
      },
...
      "dto.SuccessMessageResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "errorhandler.HttpError": {
        "description": "This error is returned for specific failures within the API",
        "properties": {
          "caller": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "message": {}
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Enter your API key in the format: x-api-key",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Enter your JWT token in the format: Bearer <token>",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "RESTful API for go-starter.",
    "title": "go-starter API",
    "version": "1.0"
  },
...
  "paths": {
    "/api/v1/auth/apple/callback": {
      "post": {
        "description": "Login with apple oauth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.CodeExchangeRequest"
              }
            }
          },
          "description": "code gotten from apple",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.CodeExchangeRequest"
                }
              }
            },
            "description": "Successfully logged in"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Login with apple",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/apple/login": {
      "get": {
        "description": "Login with apple oauth open in browser for login",
        "responses": null,
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Login with apple",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/google/callback": {
      "post": {
        "description": "Login with google oauth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.CodeExchangeRequest"
              }
            }
          },
          "description": "code gotten from google",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.CodeExchangeRequest"
                }
              }
            },
            "description": "Successfully logged in"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Login with google",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/google/login": {
      "get": {
        "description": "Login with google oauth open in browser for login",
        "responses": null,
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Login with google",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Login with email and password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.EmailLoginRequest"
              }
            }
          },
          "description": "Email Login Request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.LoginResponse"
                }
              }
            },
            "description": "Successfully logged in"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Login with email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "get": {
        "description": "invaliate current session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.SuccessMessageResponse"
                }
              }
            },
            "description": "Sign out successful."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "logout",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password-reset": {
      "post": {
        "description": "generates password reset link and sends it to the user email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.PasswordResetRequest"
              }
            }
          },
          "description": "password reset request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.PasswordResetRequest"
                }
              }
            },
            "description": "Successfully retrieved email"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "email is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "email not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get email verification link",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "description": "Resets the password for the user, token is sent in x-api-token, token and password are required",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.PasswordResetConfirmRequest"
              }
            }
          },
          "description": "password reset confirm request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.SuccessMessageResponse"
                }
              }
            },
            "description": "password reset successfully"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Token expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "User not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "x-reset-token": []
          }
        ],
        "summary": "Reset password api",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password-token-check": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.PasswordResetTokenCheck"
              }
            }
          },
          "description": "password reset confirm request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.SuccessMessageResponse"
                }
              }
            },
            "description": "password reset successfully"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Token expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "User not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "x-reset-token": []
          }
        ],
        "summary": "Reset Token check api",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/password-update": {
      "post": {
        "description": "update password from user settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.PasswordUpdateRequest"
              }
            }
          },
          "description": "password update request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.PasswordUpdateRequest"
                }
              }
            },
            "description": "Password updated successfully"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Bad requst"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update Password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh-token": {
      "get": {
        "description": "Uses a valid refresh token to generate a new access token",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.RefreshTokenResponse"
                }
              }
            },
            "description": "New access token"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid refresh token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "openapi": "3.1.0"
}